refactor: migrate cjsToEsm.cjs to esm.mjs
